Add files via upload
The dataset used in this provides detailed records of road accidents that occurred during January 2021 in Kensington and Chelsea . It includes information such as the accident date, day of the week, junction control, accident severity, lighting and weather conditions, vehicle details, and more. The code helps in visualizing the distribution of various factors . Patterns on lighting conditions during the accidents , Time of occurence of the accident , day of the accident , type of vehicle is obtained using various plots.